python_requirements.txt
P0f
python-nmap
elevate
netifaces
pyyaml
lxml
yattag
requests
get-mac
